- aggiunto il supporto per i chip miltipli di CHR ROM - aggiunta la mapper mapper_Malee - migliorato il supporto al formato UNIF

+++ src/mappers/mapper_BMCFK23C.h
@@ -19,7 +19,7 @@ struct _bmcfk23c {
 	BYTE A001;
 	BYTE reg[8];
 	BYTE mmc3[8];
-	BYTE chr_map[8];
+	WORD chr_map[8];
 	/* questo posso tranquillamente non salvarlo */
 	BYTE prg_mask;
 } bmcfk23c;
